# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Cetacean API
  version: "1.0"
  description: |
    Observability and management API for Docker Swarm Mode clusters.

    Uses content negotiation — same URLs serve JSON, HTML (SPA), or SSE based on
    the Accept header. Append `.json` or `.html` to any path to force a content type.

    All collection responses include JSON-LD `@context` and `@type` fields.
    Detail responses include `@context`, `@id`, and `@type`.
    Error responses follow RFC 9457 (Problem Details).

    Every response (except `/-/` meta endpoints) includes:
    - `Link: </api>; rel="service-desc"` — pointer to this spec
    - `Link: </api/context.jsonld>; rel="describedby"` — JSON-LD context
    - `Vary: Accept` — content negotiation indicator
servers:
  - url: /
security:
  - cookieAuth: []
  - bearerAuth: []
  - {} # Allow unauthenticated (none mode)
...
paths:
...
  /nodes:
    get:
      operationId: listNodes
      tags: [Nodes]
      summary: List nodes
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: sort
          in: query
          schema:
            type: string
            enum: [hostname, role, status, availability]
        - $ref: "#/components/parameters/dir"
      responses:
        "200":
          description: Paginated node list
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Link:
              $ref: "#/components/headers/Link"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeCollection"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
  /nodes/{id}:
    get:
      operationId: getNode
      tags: [Nodes]
      summary: Node detail
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Node detail
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      node:
                        type: object
                        description: Docker swarm.Node object
        "304":
          $ref: "#/components/responses/NotModified"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeNode
      tags: [Nodes]
      summary: Remove a node
      description: Removes a node from the swarm. The node must be drained or down. Requires operations level 3.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "204":
          description: Node removed
        "403":
          description: Operations level too low
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /nodes/{id}/availability:
    put:
      operationId: updateNodeAvailability
      tags: [Nodes]
      summary: Update node availability
      description: Sets the availability of a node (active, drain, or pause).
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [availability]
              properties:
                availability:
                  type: string
                  enum: [active, drain, pause]
      responses:
        "200":
          description: Updated node
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /nodes/{id}/labels:
    get:
      operationId: getNodeLabels
      tags: [Nodes]
      summary: Get node labels
      description: Returns the current labels of the node as a key/value map.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Node labels
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchNodeLabels
      tags: [Nodes]
      summary: Patch node labels
      description: |
        Applies a JSON Patch (RFC 6902) to the node labels.
        Requires `Content-Type: application/json-patch+json`.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        required: true
        content:
          application/json-patch+json:
            schema:
              $ref: "#/components/schemas/JSONPatch"
      responses:
        "200":
          description: Updated node labels
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /nodes/{id}/role:
    get:
      operationId: getNodeRole
      tags: [Nodes]
      summary: Get node role
      description: Returns the current role of the node (worker or manager).
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Node role
          content:
            application/json:
              schema:
                type: object
                properties:
                  role:
                    type: string
                    enum: [worker, manager]
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateNodeRole
      tags: [Nodes]
      summary: Update node role
      description: Promotes or demotes a node. Requires operations level 3.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [worker, manager]
      responses:
        "200":
          description: Updated node
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Operations level too low
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ── Services ─────────────────────────────────────────────────────────
...
  /secrets:
    get:
      operationId: listSecrets
      tags: [Secrets]
      summary: List secrets
      description: Secret data is always redacted (cleared before response).
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, created, updated]
        - $ref: "#/components/parameters/dir"
      responses:
        "200":
          description: Paginated secret list
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Link:
              $ref: "#/components/headers/Link"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretCollection"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: createSecret
      tags: [Secrets]
      summary: Create a secret
      description: Creates a new secret in the swarm. Requires operations level 2.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, data]
              properties:
                name:
                  type: string
                  description: Secret name
                data:
                  type: string
                  format: byte
                  description: Secret data (base64-encoded)
      responses:
        "201":
          description: Secret created
          headers:
            Location:
              schema:
                type: string
              description: Path to the created secret
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      secret:
                        type: object
                        description: Docker swarm.Secret object (data is always cleared)
                      services:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Operations level too low
        "409":
          description: Secret name conflict
  /secrets/{id}:
    get:
      operationId: getSecret
      tags: [Secrets]
      summary: Secret detail
      description: |
        Returns the secret metadata (data always redacted) and
        a list of services using this secret.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Secret detail with cross-referenced services
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      secret:
                        type: object
                        description: Docker swarm.Secret object (data always nil)
                      services:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRef"
        "304":
          $ref: "#/components/responses/NotModified"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeSecret
      tags: [Secrets]
      summary: Remove a secret
      description: Removes a secret from the swarm. Fails if the secret is in use by a service. Requires operations level 3.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "204":
          description: Secret removed
        "403":
          description: Operations level too low
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /secrets/{id}/labels:
    get:
      operationId: getSecretLabels
      tags: [Secrets]
      summary: Get secret labels
      description: Returns the current labels of the secret as a key/value map.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Secret labels
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchSecretLabels
      tags: [Secrets]
      summary: Patch secret labels
      description: |
        Applies a JSON Patch (RFC 6902) to the secret labels.
        Requires `Content-Type: application/json-patch+json`.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        required: true
        content:
          application/json-patch+json:
            schema:
              $ref: "#/components/schemas/JSONPatch"
      responses:
        "200":
          description: Updated secret labels
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"

  # ── Networks ─────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-cetacean_session
      description: Session cookie set after OIDC authentication
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for machine-to-machine OIDC authentication

  # ── Parameters ───────────────────────────────────────────────────────
  parameters:
    resourceId:
      name: id
      in: path
      required: true
      description: Resource ID
      schema:
        type: string
    search:
      name: search
      in: query
      description: Case-insensitive substring search on resource names
      schema:
        type: string
    filter:
      name: filter
      in: query
      description: |
        expr-lang filter expression (max 512 characters).
        Available fields depend on the resource type.
      schema:
        type: string
        maxLength: 512
    limit:
      name: limit
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    offset:
      name: offset
      in: query
      description: Starting offset for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    dir:
      name: dir
      in: query
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: asc
...
  headers:
    ETag:
      description: Weak ETag for conditional requests (use with If-None-Match)
      schema:
        type: string
        example: '"w/abc123"'
    Link:
      description: RFC 8288 pagination links (rel="next", rel="prev")
      schema:
        type: string
        example: '</nodes?limit=50&offset=50>; rel="next"'
...
  responses:
    NotModified:
      description: Not modified (ETag matched If-None-Match)
    BadRequest:
      description: Bad request (invalid filter, missing parameter, etc.)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    Conflict:
      description: Version conflict — the resource was modified concurrently; retry the operation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    UnsupportedMediaType:
      description: Unsupported Content-Type for this endpoint
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
...
    ProblemDetail:
      type: object
      description: |
        RFC 9457 Problem Details with JSON-LD @context.
        Domain-specific errors use `/api/errors/{code}` as the type URI
        (e.g. `/api/errors/SVC001`). Generic errors use `about:blank`.
      required: [type, title, status]
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        type:
          type: string
          description: "Problem type URI — `/api/errors/{code}` for domain errors, `about:blank` for generic"
          example: /api/errors/SVC001
        title:
          type: string
          description: Short human-readable summary
          example: Service Version Conflict
        status:
          type: integer
          description: HTTP status code
          example: 409
        detail:
          type: string
          description: Human-readable explanation
        instance:
          type: string
          description: Request path that generated the error
        requestId:
          type: string
          description: Correlation ID for this request
...
    DetailEnvelope:
      type: object
      description: JSON-LD wrapper for detail endpoints
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        "@id":
          type: string
          description: Canonical path of this resource
          example: /nodes/abc123
        "@type":
          type: string
          description: Resource type name
          example: Node
    CollectionEnvelope:
      type: object
      description: JSON-LD wrapper for list endpoints
      required: [items, total, limit, offset]
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        "@type":
          type: string
          example: Collection
        items:
          type: array
          items: {}
        total:
          type: integer
          description: Total number of items matching the query
        limit:
          type: integer
          description: Page size used
        offset:
          type: integer
          description: Starting offset used

    # ── Collections ──────────────────────────────────────────────────
    NodeCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                description: Docker swarm.Node object
...
    SecretCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                description: Docker swarm.Secret object (data always nil)
...
    ServiceRef:
      type: object
      description: Lightweight service cross-reference
      properties:
        "@id":
          type: string
          description: Canonical path
          example: /services/abc123
        id:
          type: string
        name:
          type: string
...
    JSONPatch:
      type: array
      description: JSON Patch document (RFC 6902)
      items:
        type: object
        required: [op, path]
        properties:
          op:
            type: string
            enum: [add, remove, replace, test]
            description: Patch operation
          path:
            type: string
            description: Key name (with or without leading /)
          value:
            type: string
            description: New value (required for add/replace/test)
...